prlmk: Rewrite to improve clarity
Performance still remains the same, though.

This is a complete rewrite, with bugfixes to the previous code changes
as well. Polluting process_reclaim code with ifdefs is not a good idea
in the long run, so this rewrite was necessary. Do note that the config
option was changed to CONFIG_PRLMK, and it isn't required to enable
CONFIG_PROCESS_RECLAIM anymore, so make the required changes to your
defconfig.

A new tunable was added, kill_heaviest_gid. This is experimental, relies
on some edge cases, and I'm not very proud of it, but it was necessary
to prevent certain classes of tasks fron being killed, like Firefox/Chrome
browser tabs for example, which usually were killed very often, even during
light multitasking. This issue also persisted across apps that launched
activities as different processes. The method used behind kill_heaviest_gid
is probably not the proper way to fix this issue, but so far its been effective.

Other tunables and the overall code retains its previous behaviors.

Signed-off-by: darkhz <kmachanwenw@gmail.com>
Change-Id: I39706e06c5429c4fd89992d31a398904a6780bc8